demo(android): unify texture-streaming + occlusion-material into materials (#2239 Batch 4)
Consolidates the retired `texture-streaming` and `occlusion-material` demos
into the existing `materials` entry with a 3-way segmented-button toggle
(mirrors #2239 Batch 1/2/3):

- **PBR Materials** (default) — streamed KHR_materials_* extension showcase.
- **Streaming** — runtime texture / material swap on a loaded SphereNode.
- **Occlusion** — invisible depth-writing occluder plane (createOcclusionInstance).

The `materials` id stays a live registered demo (the natural umbrella). Each
sub-mode owns its own rememberEngine + loaders + SceneView so switching tabs
tears down the inactive section completely (prevents resource leaks). All
QA-day / #issue rationale comments preserved from the source demos.

Retired deep-link ids (`texture-streaming`, `occlusion-material`) keep working
via DeepLinkRouter.DEMO_ID_ALIASES. Their fragments + strings folded into the
unified `materials` demo; llms.txt demos block regenerated via collate-demos.sh.

Tests: DemoInteractionTest.materials_allTabs taps all 3 tabs;
DemoRenderingScreenshotTest.materialsDemo_default_state (golden materials_default).
