+++ changelog.d/2239-batch-4-materials.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Samples catalog: unified Materials demo ([#2239](https://github.com/sceneview/sceneview/issues/2239)).** The retired `texture-streaming` and `occlusion-material` demos consolidated into the existing `materials` entry with a segmented-button toggle between PBR Materials (`KHR_materials_*` extension showcase), Streaming (runtime texture / material swap on a loaded model), and Occlusion (invisible depth-writing surface) modes. The `materials` id stays a live registered demo; existing `sceneview://demo/texture-streaming|occlusion-material` deep links keep working as aliases via `DEMO_ID_ALIASES`.
