+++ llms.txt
@@ -4019,12 +4019,10 @@ by `samples/android-demo/scripts/collate-demos.sh` — never edit between the ma
 - `custom-geometry` — Custom Geometry. Composite meshes and shape extrusion.
 - `debug-overlay` — Debug Overlay. Performance stats overlay.
 - `double-pendulum` — Double Pendulum. Chaotic two-link physics, shared KMP simulation.
-- `materials` — PBR Materials. KHR_materials_* extension showcase — sheen, transmission, iridescence.
-- `occlusion-material` — Occlusion Material. Invisible depth-writing surface hides content behind it.
+- `materials` — Materials. PBR extension showcase, runtime material streaming, and occlusion material.
 - `physics` — Physics. Gravity, collisions, and rigid bodies.
 - `secondary-camera` — Secondary Camera (PiP). Picture-in-picture camera view.
 - `spatial-audio` — Spatial Audio. 3D positional sound that pans as you orbit.
-- `texture-streaming` — Material Streaming. Swap textures and materials on a model at runtime.
 
 ### Augmented Reality
