+++ samples/android-demo/src/androidTest/java/io/github/sceneview/demo/DemoInteractionTest.kt
@@ -594,6 +594,37 @@ class DemoInteractionTest {
         screenshot("35_lab_postFx_back_to_defaults")
     }
 
+    // ── 8b. Materials — all 3 segmented tabs ──────────────────────────────────
+
+    @Test
+    fun materials_allTabs() {
+        // #2239 Batch 4 — `texture-streaming` and `occlusion-material` consolidated
+        // into the existing `materials` entry with a 3-way segmented toggle. One test
+        // taps through every tab so each merged half is exercised (the unified demo
+        // opens on its default PBR Materials tab).
+        openDemo("materials")
+
+        // ── PBR Materials tab (default landing tab) — KHR_materials_* chips ────
+        screenshot("38_materials_pbr_default")
+
+        // ── Streaming tab — runtime material-set swap on a sphere ─────────────
+        tap("Streaming")
+        screenshot("38a_materials_streaming_default")
+        tap("Copper")
+        screenshot("38b_materials_streaming_copper")
+        tap("Matte Plastic")
+        screenshot("38c_materials_streaming_plastic")
+
+        // ── Occlusion tab — invisible depth-writing occluder plane ────────────
+        // The "Occluder visible" Switch is the only interactive control; it sits in a
+        // plain Row (not a `toggleable` row like the Post-FX switches), so `tap(text)`
+        // can't reach the Switch from its sibling label. Capturing the default-state
+        // frame is enough to exercise the merged Occlusion section here; the toggle
+        // behaviour is verified visually during device-QA.
+        tap("Occlusion")
+        screenshot("38d_materials_occlusion_default")
+    }
+
     // ── 9. Debug Overlay — preset reset ───────────────────────────────────────
 
     @Test
