--- samples/android-demo/src/main/res/values/strings_demo_texture_streaming.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `texture-streaming` demo (registered by
-     `TextureStreamingFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_texture_streaming_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_texture_streaming_title">Material Streaming</string>
-    <string name="demo_texture_streaming_subtitle">Swap textures and materials on a model at runtime</string>
-    <string name="demo_texture_streaming_picker_label">Material set</string>
-    <string name="demo_texture_streaming_variant_ceramic">Glazed Ceramic</string>
-    <string name="demo_texture_streaming_variant_copper">Copper</string>
-    <string name="demo_texture_streaming_variant_detail">metallic %1$.2f · roughness %2$.2f</string>
-    <string name="demo_texture_streaming_variant_gold">Brushed Gold</string>
-    <string name="demo_texture_streaming_variant_plastic">Matte Plastic</string>
-    <string name="demo_texture_streaming_variant_steel">Polished Steel</string>
-</resources>
